requirements.txt
line-bot-sdk==3.14.0
fastapi
uvicorn[standard]
google-generativeai
pydantic
tiktoken
Pillow
aiofiles
requests